backend/RAG_Chatbot/pyproject.toml
[tool.poetry]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "RAG chatbot backend for Physical AI & Humanoid Robotics book"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
qdrant-client = "^1.7.0"
sentence-transformers = "^2.2.2"
google-generativeai = "^0.3.1"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
flake8 = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
preload = "scripts.preload:main"